gradio_app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
gradio
Pillow
numpy<2.0